requirements.txt
process-mining-core==0.1.3
apache-flink~=1.17.2
apache-flink-libraries~=1.17.2
python-dotenv~=1.0.1
pm4py~=2.7.11.13
